Fixing some compiler warnings. Added non-functional pin info to Raspberry Pi and CHIP. Get fd number using reflection rather than sun.misc.SharedSecrets so this will work on other JDKs as well as Java 9. Fixed a bug in PinInfo for SBCs with multiple headers. Fixed a bug in SystemInfo.lookupBoardInfo. PWM support in the web app sample.

+++ diozero-core/src/main/java/com/diozero/PiconZeroMotor.java
@@ -28,7 +28,6 @@ package com.diozero;
 
 import java.io.IOException;
 
-import com.diozero.PiconZero;
 import com.diozero.api.motor.MotorBase;
 import com.diozero.util.RuntimeIOException;
 
@@ -64,7 +63,7 @@ public class PiconZeroMotor extends MotorBase {
 	 */
 	@Override
 	public float getValue() throws RuntimeIOException {
-		return piconZero.getMotor(motor) / (float) Math.abs(MAX_BACKWARD_SPEED);
+		return piconZero.getMotor(motor) / Math.abs(MAX_BACKWARD_SPEED);
 	}
 
 	@Override
